Move damage classes

+++ include/PokedexEntry.h
@@ -9,7 +9,7 @@ public:
                  Type type2,
                  int base_hp, int base_atk, int base_def,
                  int base_sp_atk, int base_sp_def, int base_speed,
-                 double height, double weight,
+                 float height, float weight,
                  const std::string& description);
 
     std::string getName() const;
@@ -21,13 +21,13 @@ public:
     int getBaseSpAtk() const;
     int getBaseSpDef() const;
     int getBaseSpeed() const;
-    double getHeight() const;
-    double getWeight() const;
+    float getHeight() const;
+    float getWeight() const;
     std::string getDescription() const;
 
 private:
     const std::string name, description;
     Type type1, type2;
     const int base_hp, base_atk, base_def, base_sp_atk, base_sp_def, base_speed;
-    const double height, weight;
+    const float height, weight;
 };
